pyproject.toml
[project]
name = "airpen"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "ai-edge-torch>=0.7.0",
    "aim>=3.29.1",
    "einops>=0.8.1",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.7",
    "numpy>=2",
    "onnx==1.17.0",
    "onnx-graphsurgeon>=0.5.8",
    "onnx-tf>=1.10.0",
    "onnx2tf>=1.28.6",
    "onnxscript>=0.5.6",
    "pyserial>=3.5",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "tensorflow>=2.20.0",
    "tf-keras>=2.20.1",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
]
